ledger, graph, snowball: initial work on snowball running in parallel to the view-graph tx, graph: have each transaction keep in-memory the "seed", or otherwise the number of prefixed zeroes of BLAKE2b(sender || parents) to make computing if a transaction is critical cheaper, and have graphs index transactions by their seed snowball: port snowball back into wavelet pkg, still needs to be updated to come to consensus on Rounds rather than Transactions tx: have difficulty represented as a byte, as the max difficulty is 255 tx/test: make code for generating random transactions simpler graph: have graph keep track of root id

+++ tx_test.go
@@ -7,14 +7,14 @@ import (
 )
 
 func TestTransaction_ExpectedDifficulty(t *testing.T) {
-	const MinDifficulty = uint64(8)
+	const MinDifficulty = byte(8)
 
-	assert.Equal(t, uint64(9), Transaction{Depth: 54638, Confidence: 196398}.ExpectedDifficulty(MinDifficulty))
+	assert.Equal(t, byte(9), Transaction{Depth: 54638, Confidence: 196398}.ExpectedDifficulty(MinDifficulty))
 
 	// Genesis transaction.
 	assert.Equal(t, MinDifficulty, Transaction{Depth: 0, Confidence: 0}.ExpectedDifficulty(MinDifficulty))
 
 	// Test upper bounds.
-	assert.Equal(t, uint64(34), Transaction{Depth: 19638, Confidence: math.MaxUint64}.ExpectedDifficulty(MinDifficulty))
+	assert.Equal(t, byte(34), Transaction{Depth: 19638, Confidence: math.MaxUint64}.ExpectedDifficulty(MinDifficulty))
 	assert.Equal(t, MinDifficulty, Transaction{Depth: math.MaxUint64, Confidence: 19638}.ExpectedDifficulty(MinDifficulty))
 }
